2015-09-28 Fred Gleason <fredg@paravelsystems.com> 	* Fixed a bug in 'lib/rdlog_event.cpp' that generated incorrect 	over/under time estimates for block timescaling. 	* Added a 'TS Block' counter to RDLogEdit's 'Edit Log' dialog in 	'rdlogedit/edit_log.cpp' and 'rdlogedit/edit_log.h'.

+++ lib/rdlog_event.cpp
@@ -529,7 +529,7 @@ double RDLogEvent::blockTimescaleRatio(int *err_msecs,int from_line,
     *err_msecs=0;
   }
   else {
-    *err_msecs=(double)(desired_len-actual_len)/ret;
+    *err_msecs=desired_len-actual_len/ret;
   }
 
   return ret;
